chore: refactoring
didn't touch the logic. Just tried to improve the structure
