+++ apps/web/app/composables/useRichTextEditor/useLinkModal/index.ts
@@ -1,0 +1,1 @@
+export { useLinkModal } from './useLinkModal';
