+++ apps/web/app/composables/useRichTextEditor/useLinkModal/useLinkModal.ts
@@ -1,62 +1,15 @@
 import type { Editor } from '@tiptap/core';
-import type { Mark as PmMark, Node as PmNode } from '@tiptap/pm/model';
+import type { Node as PmNode } from '@tiptap/pm/model';
 import { NodeSelection } from '@tiptap/pm/state';
-import { validateUrl } from '~/composables/useRichTextEditor/helpers/url.helper';
-import type { LinkModalLinkAttrs } from '~/composables/useRichTextEditor/types';
-
-const rangeContainsInlineAtoms = (doc: PmNode, from: number, to: number): boolean => {
-  let found = false;
-  doc.nodesBetween(from, to, (n) => {
-    if (!found && n.isAtom && n.isInline) found = true;
-    return !found;
-  });
-  return found;
-};
-
-const isAtomNode = (n: PmNode) => n.type.name === 'icon' || n.type.name === 'emoji';
-
-const rangeContainsAtoms = (doc: PmNode, from: number, to: number) => {
-  let found = false;
-  doc.nodesBetween(from, to, (n) => {
-    if (!found && isAtomNode(n)) {
-      found = true;
-    }
-    return !found;
-  });
-  return found;
-};
-
-const getLinkMark = (editor: Editor, from: number, to: number): PmMark | undefined => {
-  if (from !== to) {
-    let found: PmMark | undefined;
-    editor.state.doc.nodesBetween(from, to, (n) => {
-      if (!found) found = n.marks.find((m) => m.type.name === 'link');
-      return !found;
-    });
-    if (found) return found;
-  }
-  const sel = editor.state.selection;
-  const marks = [...(editor.state.storedMarks ?? []), ...sel.$from.marks()];
-  return marks.find((m) => m.type.name === 'link');
-};
-
-const getLinkAttrsFromRange = (doc: PmNode, from: number, to: number) => {
-  let result: LinkModalLinkAttrs | null = null;
-  doc.nodesBetween(from, to, (n) => {
-    if (result) return false;
-    const m = n.marks.find((m) => m.type.name === 'link');
-    if (m) {
-      result = {
-        href: m.attrs.href as string,
-        target: m.attrs.target as string,
-        'data-link-type': m.attrs['data-link-type'] as LinkModalLinkAttrs['data-link-type'],
-        'data-link-value': m.attrs['data-link-value'] as string | undefined,
-        'data-link-path': m.attrs['data-link-path'] as string | undefined,
-      };
-    }
-  });
-  return result;
-};
+import { validateUrl, isCategoryHref, isStaticPageHref } from './url.helper';
+import {
+  rangeContainsInlineAtoms,
+  rangeContainsAtoms,
+  getLinkMark,
+  getLinkAttrsFromRange,
+  isAtomNode,
+} from './helpers';
+import type { LinkModalLinkAttrs } from '../types';
 
 export const useLinkModal = (editor: Ref<Editor | null | undefined> | ComputedRef<Editor | null | undefined>) => {
   const tabs: { value: LinkTabValue; label: string }[] = [
@@ -78,10 +31,11 @@ export const useLinkModal = (editor: Ref<Editor | null | undefined> | ComputedRe
   const isInitializing = ref(false);
   const originalAtomLinkFullAttrs = ref<({ href: string; target: string } & LinkModalLinkAttrs) | null>(null);
   const atomDisplayLabel = ref<string | null>(null);
+  const initialLinkAttrs = ref<LinkModalLinkAttrs | null>(null);
+  const initialSelection = ref<{ from: number; to: number } | null>(null);
+  const initialText = ref('');
 
   const { getCorrectPreviewPathWithLocale } = useCategoryIdHelper();
-
-  const initialLinkAttrs = ref<LinkModalLinkAttrs | null>(null);
 
   const categoryHref = computed(() => {
     if (!selectedCategoryPath.value) return '';
@@ -93,17 +47,6 @@ export const useLinkModal = (editor: Ref<Editor | null | undefined> | ComputedRe
     const { from, to } = editor.value.state.selection;
     return editor.value.state.doc.textBetween(from, to, ' ');
   });
-
-  const initialSelection = ref<{ from: number; to: number } | null>(null);
-  const initialText = ref('');
-
-  const isCategoryHref = (href: string) => {
-    return href.includes('/c/') || href.includes('/category/');
-  };
-
-  const isStaticPageHref = (href: string) => {
-    return href.startsWith('/') && !isCategoryHref(href);
-  };
 
   const extractLinkAttrs = (rawAttrs?: LinkModalLinkAttrs): LinkModalLinkAttrs => ({
     href: rawAttrs?.href,
@@ -356,6 +299,7 @@ export const useLinkModal = (editor: Ref<Editor | null | undefined> | ComputedRe
 
     onClose();
   };
+
   return {
     activeTab,
     urlValue,
